pyproject.toml
[project]
name = "hackthebot"
version = "0.1.0"
description = "Bot of HackTheBox"
authors = [
    {name = "BlaiseBits"}
]
license = {text = "to kill"}
readme = "README.md"
requires-python = ">=3.12,<4.0"
dependencies = [
    "langchain (>=0.3.14,<0.4.0)",
    "langchain-openai (>=0.3.0,<0.4.0)",
    "langchain-community (>=0.3.14,<0.4.0)",
    "python-libnmap (>=0.7.3,<0.8.0)",
    "python-dotenv (>=1.0.1,<2.0.0)",
    "langchain-anthropic (>=0.3.8,<0.4.0)",
    "argparse (>=1.4.0,<2.0.0)",
    "langgraph (>=0.3.5,<0.4.0)",
    "langgraph-cli[inmem] (>=0.1.77,<0.2.0)",
    "retry (>=0.9.2,<0.10.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
